# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gin-admin-api 后台管理接口文档",
        "title": "gin-admin-api",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/role": {
            "post": {
                "description": "创建后台角色，并分配角色资源关系",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色中心"
                ],
                "summary": "创建角色",
                "parameters": [
                    {
                        "description": "创建角色参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gin-admin-api_internal_api_role_dto.CreateSysRoleDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/role/detail/{id}": {
            "get": {
                "description": "根据角色 ID 获取角色详情和已授权资源 ID 列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色中心"
                ],
                "summary": "获取角色详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "角色详情",
                        "schema": {
                            "$ref": "#/definitions/vo.SysRoleDetailVO"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/role/page": {
            "post": {
                "description": "根据查询条件分页获取角色列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色中心"
                ],
                "summary": "分页获取角色",
                "parameters": [
                    {
                        "description": "分页查询参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RolePageDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "角色分页列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/gin-admin-api_internal_api_role_vo.SysRoleVO"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/role/{id}": {
            "put": {
                "description": "根据角色 ID 修改角色基础信息和角色资源关系",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色中心"
                ],
                "summary": "修改角色",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改角色参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gin-admin-api_internal_api_role_dto.CreateSysRoleDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据角色 ID 删除角色，并清理角色资源关系",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色中心"
                ],
                "summary": "删除角色",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.RolePageDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "status": {
                    "description": "状态1是正常,2是禁用",
                    "type": "integer"
                }
            }
        },
...
        "gin-admin-api_internal_api_role_dto.CreateSysRoleDTO": {
            "type": "object",
            "required": [
                "dataScope",
                "name",
                "resourcesIdList"
            ],
            "properties": {
                "dataScope": {
                    "type": "integer",
                    "enum": [
                        1,
                        2,
                        3,
                        4,
                        5
                    ]
                },
                "deptIdList": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "resourcesIdList": {
                    "description": "资源id",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "integer"
                    }
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态1是正常,2是禁用",
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                }
            }
        },
        "gin-admin-api_internal_api_role_vo.SysRoleVO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "dataScope": {
                    "description": "数据范围",
                    "type": "integer"
                },
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "id": {
                    "description": "主键id",
                    "type": "integer"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态1是正常,2是禁用",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "integer"
                }
            }
        },
...
        "vo.SysRoleDetailVO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "dataScope": {
                    "description": "数据范围",
                    "type": "integer"
                },
                "deptIdList": {
                    "description": "自定义数据权限部门id",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "id": {
                    "description": "主键id",
                    "type": "integer"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "resourcesIdList": {
                    "description": "授权的资源id",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态1是正常,2是禁用",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "integer"
                }
            }
        }
    }
}